+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## v8.4.0 (2024-03-22)
+
+### Feat
+
+- Add Discharge Reciprocating Pump function
+
+## v8.3.3 (2024-03-20)
+
+### Fix
+
+- change  descriptions of arguments
+- **deps**: update dependency numba to ^0.59.0
+
 ## v8.3.2 (2024-02-23)
 
 ### Fix